app/microservices/smsh_ia_models/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.14
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
pytorch==2.6.0
transformers==4.36.2
